% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{calcular_diferencia}
\alias{calcular_diferencia}
\title{Calcula la diferencia entre dos tiempos dependiendo del estatus}
\usage{
calcular_diferencia(bd, fecha_reciente, fecha_antigua, unidades = "hours")
}
\arguments{
\item{bd}{base de datos}

\item{fecha_reciente}{fecha más reciente dependiendo del estatus}

\item{fecha_antigua}{fecha más antigua dependiendo del estatus}

\item{unidades}{unidades de tiempo, por default es 'hours'.}

\item{estatus}{estatus de la solicitud}
}
\description{
Calcula la diferencia entre dos tiempos dependiendo del estatus.
También se calcula la media, el mínimo y el máximo, por si esta información es de utilidad.
}
\examples{
data <- calcular_diferencia(ex_data, "finalizada", "fecha_atendida", "fecha")
}
