% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{graficar_boxplot}
\alias{graficar_boxplot}
\title{Graficar boxplot}
\usage{
graficar_boxplot(bd, var, nombre, color, grupo, titulo, eje_x, eje_y)
}
\arguments{
\item{bd}{Base de datos con una columna 'dif'}

\item{var}{Variable x, normalmente es 'dif'}

\item{nombre}{Nombre de la variable}

\item{color}{Color de la gráfica}

\item{grupo}{Variable y, normalmente es 'estatus'}

\item{titulo}{Título de la gráfica}

\item{eje_x}{Eje x de la gráfica}

\item{eje_y}{Eje y de la gráfica}
}
\value{
Boxplot
}
\description{
Grafica un boxplot
}
\examples{
g <- graficar_boxplot(data, var = "dif", nombre = "Tiempo de espera", color = "#2a9d8f", grupo = "estatus", titulo = "Tiempo de espera de las solicitudes finalizadas", eje_x = "Tiempo de espera (horas)", eje_y = "Estatus")
}
