% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{graficar_rainclouds}
\alias{graficar_rainclouds}
\title{Graficar rainclouds}
\usage{
graficar_rainclouds(bd, x, y, color, titulo, eje_x)
}
\arguments{
\item{bd}{Base de datos con una columna 'dif'}

\item{x}{Variable x, normalmente es 'dif'}

\item{y}{Variable y, normalmente es 'estatus'}

\item{color}{Color de la gráfica}

\item{titulo}{Título de la gráfica}

\item{eje_x}{Eje x de la gráfica}
}
\value{
Raincloud plot
}
\description{
Grafica rainclouds, las rainclouds son la combinación de un boxplot, un punto y una densidad.
}
\examples{
g <- graficar_rainclouds(data, x = "dif", y = "estatus", color = "#2a9d8f", titulo = "Tiempo de espera de las solicitudes finalizadas", eje_x = "Tiempo de espera (horas)")
}
